Create Chat log Written with code and checked at the bottom
This is basically just me copying and pasting the chat log, and then the updated code the log provided. I then went on to use to the code with the data provided from my loan dataset. The bottom were the results of my Copilot codes, that answers the Question "What were the 3 most common Loan interest rates in this dataset."